Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,9 +2,9 @@
 #include<math.h>
 int main()
 {
-     double  int a,b,c=2;
+    int a,b,c=2;
     scanf("%d %d",&a,&b);
-    a = pow(b,c);
+    (int)a = pow(b,c);
     if(a=b)
     {
         printf("Yes");
